pyproject.toml
[project]
name = "plank-road"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "APScheduler",
  "ariadne-split==0.1.0",
  "grpcio-tools==1.78.1",
  "grpcio==1.78.1",
  "huggingface_hub==1.5.0",
  "imutils",
  "loguru==0.7.3",
  "mapcalc",
  "numpy==2.2.6",
  "opencv-python==4.11.0.86",
  "ortools==9.14.6206",
  "pandas==2.3.3",
  "protobuf==6.31.1",
  "psutil==7.2.2",
  "pytest==9.0.2",
  "PyYAML==6.0.3",
  "requests==2.32.5",
  "rfdetr>=1.6.5",
  "timm>=1.0.26",
  "torch>=2.11.0",
  "torchlens==1.0.1",
  "torchvision>=0.26.0",
  "transformers",
  "ultralytics>=8.4.41",
]

[project.optional-dependencies]
integration = [
  "rfdetr>=1.6.5",
  "timm>=1.0.26",
  "torchvision>=0.26.0",
  "ultralytics>=8.4.41",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
  "ignore:`torch.jit.script` is deprecated.*:DeprecationWarning:torch.jit._script",
  "ignore:rfdetr\\.util\\..* is deprecated.*:DeprecationWarning",
]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [
  ".venv",
  "__pycache__",
  "grpc_server/message_transmission_pb2.py",
  "grpc_server/message_transmission_pb2_grpc.py",
  "grpc_server/protos/message_transmission_pb2.py",
  "grpc_server/protos/message_transmission_pb2_grpc.py",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402"]
